Fixed the IDmaps dictionary and updated docstrings.
The previous IDmaps were a nested dictionary with the master-key given by the snapshot field. This isn't correct because we want a global dictionary so that the merger pointers can be updated properly.
Updating this now-global dictionary takes a lot of time.  A possible solution is to use a nested dictionary master-keyed by the SnapNum.  Then when we need to grab a merger pointer, first turn the ID into a SnapNum then access the correct sub-dictionary.