Changed IDmaps to be a nested dictionary.
Now when we change the oldID to newID we determine what SnapNum the ID corresponds to (doesn't need to be the current SnapNum, e.g., 'Head' and 'Tail' almost certainly won't be) and then use that to key the nested IDmaps dictionary.  IDmaps is no longer one large dictionary that needs to be constantly appended to.  Runtime for creating the dictionary has gone from ~80 seconds to ~7 seconds.

Doing a few more checks to make sure outputs are identical.