Finished testing script.
This testing script tests the output of the forest_sorter code.  If the user specifies their own
input data, the script will grab the first N halos (default 10,000 but can be specified by user)
and write them to a temporary hdf5 file.  If they don't specify their own data we'll use a default
testing data set (to be added to git as it'll only be a few MB).

Once this data has been acquired, we run the forest_sorter algorithm and save the output.  We then
go through the output and check that the IDs are temporally unique and match the formula provided,
that the order of ForestID and Mass_200mean (or whatever keys request to sort on) have been saved
correctly and finally that the order of the non-ID properties were saved in the correct order.