Changed from optparse to argparse.
Optparse is apparently deprecated as of Python2.7 so moved to new argparse.
The API is basically identical however.

Also I know handle the user specification of the sort_fields and ID_fields as a
single string.  So if the user wants to sort on 'ForestID' and 'NextSubhalo'
they would specify it as 'python forest_sorter.py
--sort_fields=ForestID,NextSubhalo'.  I then split the string.