Work on fixing flybys (#1).
This commit implements a function to fix the flybys.  Currently this solution
runs fully (i.e., no key errors) but again the runtime is exceedingly long (~8
hours).

I am unsure if this can be avoided.  Profiling the code, ~65% of the runtime
occurs on the ``for`` loop over the IDmaps dictionary.  This is unavoidable.