Broke flybys (#1) but it's ok!
The old runtime was simply unacceptable.  For some reason, I was trying to run
the previous commit and it was slowing down EVEN further (like, taking 10
SECONDS to do one forest out of 500,000!!!!).  So I decided to rework the flow.

In this commit, the IDmaps are built globally for all forests.  Then we loop
over the snapshots a final time and update the IDs one and for all.  This way I
am not being killed by time taken to loop through dictionaries or slice through
the HDF5 array.

The current implementation of flybys does not work because that was written
assuming a single forest was being passed.  This could still work but I'll need
to think more on it.

This commit also lacks proper documentation for my updates because it's home
time...